Se revisan rutas dy funciones de los mantenedores en general, se mejoran validaciones tanto al crear, modificar y eliminar registros

+++ controllers/institucionController.js
@@ -1,4 +1,4 @@
-const { Institucion, Usuario } = require('../config/db');
+const { Institucion, Curso } = require('../config/db');
 const { Sequelize, Op } = require('sequelize');
 const { validationResult } = require('express-validator');
 
@@ -6,65 +6,28 @@ exports.crearInstitucion = async(req, res) => {
 
     //si hay errores de la validación
     const errors = validationResult(req);
+    console.log(errors);
     if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+        return res.status(400).json({ errors: errors.array({ onlyFirstError: true }) });
     }
 
     try {
 
-        const { codigo, descripcion, rut_usuario_rector, rut_usuario_administrador,direccion,email, telefono,
+        let { codigo, descripcion, direccion, email, telefono,
                 website ,logo, inactivo } = req.body;
 
         //verifica que las institucion no existe.
         let institucion = await Institucion.findByPk(codigo);
         if (institucion) {
-            console.log('La institución ya existe');
             return res.status(400).json({
                 msg: 'La institución ya existe'
             });
         }
 
-        //Verifica que el rut_usuario_rector sea valido y que corresponda a un rector.
-        let usuario_rector = await Usuario.findByPk(rut_usuario_rector);
-        if (usuario_rector) {
-            if (usuario_rector.dataValues.codigo_rol != 4) {
-                console.log('El rut ingresado no corresponde a un rector');
-                return res.status(400).json({
-                    msg: 'El rut ingresado no corresponde a un rector'
-                });
-            }
-        } else {
-            console.log('El rector ingresado no es válido');
-            return res.status(400).json({
-                msg: 'El rector ingresado no es válido'
-            });
-        }
-
-        //Verifica que el rut_usuario_administrador sea valido y corresponda a un administrador
-        let usuario_administrador = await Usuario.findByPk(rut_usuario_administrador);
-        if (usuario_administrador) {
-            if (usuario_administrador.dataValues.codigo_rol != 5) {
-                console.log('El rut ingresado no corresponde a un administrador');
-                return res.status(400).json({
-                    msg: 'El rut ingresado no corresponde a un administrador'
-                });
-            }
-
-        } else {
-            console.log('El administrador ingresado no es válido');
-            return res.status(400).json({
-                msg: 'El administrador ingresado no es válido'
-            });
-        }
-
-
-
         //Guarda la nueva institucion
         institucion = await Institucion.create({
             codigo,
             descripcion,
-            rut_usuario_rector,
-            rut_usuario_administrador,
             direccion,
             email,
             telefono,
@@ -74,7 +37,9 @@ exports.crearInstitucion = async(req, res) => {
         });
 
         //envía la respuesta
-        res.json(institucion);
+        res.json({
+            institucion
+        });
 
     } catch (error) {
         console.log(error);
@@ -104,14 +69,16 @@ exports.listarInstituciones = async(req, res) => {
 
 exports.actualizarInstitucion = async(req, res) => {
 
+    //si hay errores de la validación
     const errors = validationResult(req);
+    console.log(errors);
     if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+        return res.status(400).json({ errors: errors.array({ onlyFirstError: true }) });
     }
 
     try {
 
-        let { codigo, descripcion, rut_usuario_rector, rut_usuario_administrador,direccion,email, telefono,
+        let { codigo, descripcion, direccion, email, telefono,
             website ,logo, inactivo } = req.body;
 
         //verifica que la institucion a actualizar existe.
@@ -125,8 +92,6 @@ exports.actualizarInstitucion = async(req, res) => {
         //actualiza los datos.
         institucion = await Institucion.update({
             descripcion,
-            rut_usuario_rector,
-            rut_usuario_administrador,
             direccion,
             email,
             telefono,
@@ -139,7 +104,7 @@ exports.actualizarInstitucion = async(req, res) => {
             }
         })
 
-        res.json(institucion);
+        res.json({institucion});
 
     } catch (error) {
         console.log(error);
@@ -148,11 +113,10 @@ exports.actualizarInstitucion = async(req, res) => {
         })
     }
 
-
 }
 
 exports.eliminarInstitucion = async(req, res) => {
-
+    
     try {
         //obtengo el codigo del request
         const { codigo } = req.params;
@@ -163,6 +127,20 @@ exports.eliminarInstitucion = async(req, res) => {
                 msg: `La institución ${codigo} no existe`
             })
         }
+
+        //verifica que no tiene cursos asociados.
+        let institucion_cursos = await Curso.findOne({
+            where: {
+                codigo_institucion: codigo
+            }
+        })
+
+        if(institucion_cursos){
+            return res.status(404).send({
+                msg: `La institución tiene cursos asociados, no se puede eliminar`
+            })
+        }
+
         //elimino el registro.
         institucion = await Institucion.destroy({
             where: {
@@ -214,13 +192,14 @@ exports.busquedaInstituciones = async(req, res) => {
 
     try {
         //obtiene el parametro desde la url
-        const { filtro } = req.params
-        console.log(filtro)
-            //consulta por la institucion
+        const { filtro } = req.query
+        //consulta por la institucion
         const instituciones = await Institucion.findAll({
-            where: Sequelize.where(Sequelize.fn("concat", Sequelize.col("codigo"), Sequelize.col("descripcion")), {
-                [Op.like]: `%${filtro}%`
-            })
+            where: 
+                Sequelize.where(Sequelize.fn("concat", Sequelize.col("codigo"), Sequelize.col("descripcion")), { [Op.like]: `%${filtro}%`}), 
+                order: [
+                    ['descripcion', 'ASC'],
+                ] 
         });
 
         //envia la información del institucion
