+++ controllers/moduloContenidoController.js
@@ -1,19 +1,26 @@
-const { ModuloContenido, Modulo, Unidad, ModuloContenidoTema } = require('../config/db');
-const { Sequelize, Op } = require('sequelize');
+const { ModuloContenido, Modulo, Unidad, ModuloContenidoTema, PreguntaModuloContenido } = require('../config/db');
+const { Op } = require('sequelize');
 const { validationResult } = require('express-validator');
 
 
-exports.crearModuloContenido = async(req, res, next) => {
+exports.crearModuloContenido = async(req, res) => {
 
     //si hay errores de la validación
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+        return res.status(400).json({ errors: errors.array({ onlyFirstError: true }) });
     }
 
     try {
       
         const { codigo, descripcion, codigo_modulo, inactivo } = req.body;
+
+        let modulo = await Modulo.findByPk(codigo_modulo);
+        if (!modulo) {
+            return res.status(404).send({
+                msg: `El modulo ${codigo_modulo} no existe`
+            })
+        }
 
         //Guarda la nueva relacion entre contenido y modulo
         modulo_contenido = await ModuloContenido.create({
@@ -23,7 +30,9 @@ exports.crearModuloContenido = async(req, res, next) => {
             inactivo
         });
 
-        res.json(modulo_contenido)
+        res.json({
+            modulo_contenido
+        })
 
 
     } catch (error) {
@@ -64,7 +73,7 @@ exports.actualizarModuloContenido = async(req, res) => {
     //si hay errores de la validación
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+        return res.status(400).json({ errors: errors.array({ onlyFirstError: true }) });
     }
 
     try {
@@ -78,12 +87,12 @@ exports.actualizarModuloContenido = async(req, res) => {
             })
         }
 
-        /* let modulo = await Modulo.findByPk(codigo_modulo);
+        let modulo = await Modulo.findByPk(codigo_modulo);
         if (!modulo) {
             return res.status(404).send({
                 msg: `El modulo ${codigo_modulo} no existe`
             })
-        } */
+        }
 
         contenido = await ModuloContenido.update({
             descripcion,
@@ -109,7 +118,7 @@ exports.actualizarModuloContenido = async(req, res) => {
 
 }
 
-exports.eliminarModuloContenido = async(req, res, next) => {
+exports.eliminarModuloContenido = async(req, res) => {
 
     //si hay errores de la validación
     const errors = validationResult(req);
@@ -129,7 +138,8 @@ exports.eliminarModuloContenido = async(req, res, next) => {
             })
         }
 
-        let temas_contenido = await ModuloContenidoTema.findAll({
+        //revisa si tiene temas asociados.
+        let temas_contenido = await ModuloContenidoTema.findOne({
             where:{
                 codigo_modulo_contenido : codigo
             }
@@ -140,6 +150,18 @@ exports.eliminarModuloContenido = async(req, res, next) => {
             });
         }
 
+        //revisa si tiene preguntas asociadas.
+        let preguntas_contenido = await PreguntaModuloContenido.findOne({
+            where:{
+                codigo_modulo_contenido : codigo
+            }
+        })
+        if (preguntas_contenido){
+            return res.status(404).send({
+                msg: `El contenido ${codigo} tiene preguntas asociadas, no se puede eliminar`
+            });
+        }
+
         //elimino el registro.
         await ModuloContenido.destroy({
             where: {
@@ -147,9 +169,9 @@ exports.eliminarModuloContenido = async(req, res, next) => {
             }
         });
 
-        //next para pasar a listarModuloContenidos
-        req.params.codigo_modulo = codigo_modulo;
-        next();
+        res.json({
+            msg: 'Contenido eliminado correctamente'
+        });
 
     } catch (error) {
         console.log(error);
@@ -160,12 +182,11 @@ exports.eliminarModuloContenido = async(req, res, next) => {
 
 }
 
-
-exports.contenidoModulo = async(req, res) => {
+exports.contenidosPorDescripcionModuloUnidadyMateria = async(req, res) => {
 
     try {
 
-        let { codigo_modulo, codigo_unidad, codigo_materia } = req.params;
+        let { codigo_modulo, codigo_unidad, codigo_materia, descripcion } = req.query;
 
         if(codigo_modulo.trim() === '0'){
             codigo_modulo = ''
@@ -178,7 +199,6 @@ exports.contenidoModulo = async(req, res) => {
         if(codigo_materia.trim() === '0'){
             codigo_materia = ''
         }
-    
 
         const modulo_contenido = await ModuloContenido.findAll({
             include:[{
@@ -191,22 +211,16 @@ exports.contenidoModulo = async(req, res) => {
             }],
             where: {
                 [Op.and]:[
+                    { descripcion: { [Op.like] : `%${descripcion}%`} },
                     { codigo_modulo: { [Op.like] : `%${codigo_modulo}%`} },
                     {'$modulo.codigo_unidad$': { [Op.like]: `%${codigo_unidad}%` } },
                     {'$modulo.unidad.codigo_materia$': { [Op.like]: `%${codigo_materia}%` } },
-                    {inactivo: false}
                 ]
             },
             order: [
                 ['descripcion', 'ASC']
             ]
         });
-
-        if (!modulo_contenido) {
-            return res.status(404).send({
-                msg: `El contenido ${codigo_modulo} no tiene modulos asociados`
-            });
-        }
 
         res.json({
             modulo_contenido
